--- angular/.angular/cache/17.3.17/babel-webpack/89c55898d0721660c4f0e1febc2317a4b2b4df8921e693a40d82ae9a79dd4d1b.json
@@ -1,1 +1,0 @@
-{"ast":null,"code":"import { __decorate } from \"tslib\";\nimport { Component } from '@angular/core';\nimport { CommonModule } from '@angular/common';\nimport { StudyHeaderComponent } from './components/study-header/study-header.component';\nimport { StudyControlsComponent } from './components/study-controls/study-controls.component';\nimport { StudyTableComponent } from './components/study-table/study-table.component';\nimport { ValidationPanelComponent } from './components/validation-panel/validation-panel.component';\nlet AppComponent = class AppComponent {\n  constructor() {\n    this.title = 'BDD Agentic Workflow';\n  }\n};\nAppComponent = __decorate([Component({\n  selector: 'app-root',\n  standalone: true,\n  imports: [CommonModule, StudyHeaderComponent, StudyControlsComponent, StudyTableComponent, ValidationPanelComponent],\n  template: `\n    <div class=\"min-h-screen bg-background\">\n      <app-study-header></app-study-header>\n      <div class=\"container mx-auto px-6 py-8 space-y-8\">\n        <app-validation-panel></app-validation-panel>\n        <div class=\"space-y-4\">\n          <div class=\"flex items-center justify-between\">\n            <h2>Study Entries</h2>\n          </div>\n          <app-study-controls></app-study-controls>\n          <app-study-table></app-study-table>\n        </div>\n      </div>\n    </div>\n  `,\n  styles: [`\n    .min-h-screen { min-height: 100vh; }\n    .bg-background { background-color: var(--background, #ffffff); }\n    .container { max-width: 1200px; }\n    .mx-auto { margin-left: auto; margin-right: auto; }\n    .px-6 { padding-left: 1.5rem; padding-right: 1.5rem; }\n    .py-8 { padding-top: 2rem; padding-bottom: 2rem; }\n    .space-y-8 > * + * { margin-top: 2rem; }\n    .space-y-4 > * + * { margin-top: 1rem; }\n    .flex { display: flex; }\n    .items-center { align-items: center; }\n    .justify-between { justify-content: space-between; }\n  `]\n})], AppComponent);\nexport { AppComponent };","map":{"version":3,"names":["Component","CommonModule","StudyHeaderComponent","StudyControlsComponent","StudyTableComponent","ValidationPanelComponent","AppComponent","constructor","title","__decorate","selector","standalone","imports","template","styles"],"sources":["F:\\aaa\\jira_2_bdd_agentic_workflow\\angular\\src\\app\\app.component.ts"],"sourcesContent":["import { Component } from '@angular/core';\nimport { CommonModule } from '@angular/common';\nimport { StudyHeaderComponent } from './components/study-header/study-header.component';\nimport { StudyControlsComponent } from './components/study-controls/study-controls.component';\nimport { StudyTableComponent } from './components/study-table/study-table.component';\nimport { ValidationPanelComponent } from './components/validation-panel/validation-panel.component';\n\n@Component({\n  selector: 'app-root',\n  standalone: true,\n  imports: [\n    CommonModule,\n    StudyHeaderComponent,\n    StudyControlsComponent,\n    StudyTableComponent,\n    ValidationPanelComponent\n  ],\n  template: `\n    <div class=\"min-h-screen bg-background\">\n      <app-study-header></app-study-header>\n      <div class=\"container mx-auto px-6 py-8 space-y-8\">\n        <app-validation-panel></app-validation-panel>\n        <div class=\"space-y-4\">\n          <div class=\"flex items-center justify-between\">\n            <h2>Study Entries</h2>\n          </div>\n          <app-study-controls></app-study-controls>\n          <app-study-table></app-study-table>\n        </div>\n      </div>\n    </div>\n  `,\n  styles: [`\n    .min-h-screen { min-height: 100vh; }\n    .bg-background { background-color: var(--background, #ffffff); }\n    .container { max-width: 1200px; }\n    .mx-auto { margin-left: auto; margin-right: auto; }\n    .px-6 { padding-left: 1.5rem; padding-right: 1.5rem; }\n    .py-8 { padding-top: 2rem; padding-bottom: 2rem; }\n    .space-y-8 > * + * { margin-top: 2rem; }\n    .space-y-4 > * + * { margin-top: 1rem; }\n    .flex { display: flex; }\n    .items-center { align-items: center; }\n    .justify-between { justify-content: space-between; }\n  `]\n})\nexport class AppComponent {\n  title = 'BDD Agentic Workflow';\n}"],"mappings":";AAAA,SAASA,SAAS,QAAQ,eAAe;AACzC,SAASC,YAAY,QAAQ,iBAAiB;AAC9C,SAASC,oBAAoB,QAAQ,kDAAkD;AACvF,SAASC,sBAAsB,QAAQ,sDAAsD;AAC7F,SAASC,mBAAmB,QAAQ,gDAAgD;AACpF,SAASC,wBAAwB,QAAQ,0DAA0D;AAyC5F,IAAMC,YAAY,GAAlB,MAAMA,YAAY;EAAlBC,YAAA;IACL,KAAAC,KAAK,GAAG,sBAAsB;EAChC;CAAC;AAFYF,YAAY,GAAAG,UAAA,EAvCxBT,SAAS,CAAC;EACTU,QAAQ,EAAE,UAAU;EACpBC,UAAU,EAAE,IAAI;EAChBC,OAAO,EAAE,CACPX,YAAY,EACZC,oBAAoB,EACpBC,sBAAsB,EACtBC,mBAAmB,EACnBC,wBAAwB,CACzB;EACDQ,QAAQ,EAAE;;;;;;;;;;;;;;GAcT;EACDC,MAAM,EAAE,CAAC;;;;;;;;;;;;GAYR;CACF,CAAC,C,EACWR,YAAY,CAExB","ignoreList":[]},"metadata":{},"sourceType":"module","externalDependencies":[]}
