--- angular/PROJECT_STRUCTURE.txt
@@ -1,78 +1,0 @@
-BDD AGENTIC WORKFLOW - ANGULAR PROJECT STRUCTURE
-=================================================
-
-bdd-agentic-workflow-angular/
-├── README.md                          # Project documentation and setup instructions
-├── package.json                       # NPM dependencies and scripts
-├── angular.json                       # Angular CLI configuration
-├── tsconfig.json                      # TypeScript configuration
-├── tsconfig.app.json                  # Application-specific TypeScript config
-├── .gitignore                         # Git ignore rules
-├── setup.bat                          # Windows setup script
-├── setup.sh                           # Unix/Linux setup script
-│
-├── src/
-│   ├── index.html                     # Main HTML file
-│   ├── main.ts                        # Application bootstrap
-│   ├── styles.scss                    # Global styles with design system
-│   ├── favicon.ico                    # Application icon
-│   │
-│   ├── assets/                        # Static assets directory
-│   │   └── .gitkeep                   # Placeholder file
-│   │
-│   └── app/
-│       ├── app.component.ts           # Root application component
-│       │
-│       └── components/
-│           ├── job-status-panel/
-│           │   └── job-status-panel.component.ts  # Combined job status panel with grid
-│           │
-│           └── feature-generation-job/
-│               └── feature-generation-job.component.ts  # Feature generation job panel and quick actions
-
-QUICK START GUIDE:
-==================
-
-1. Prerequisites:
-   - Node.js (v16 or higher)
-   - npm (comes with Node.js)
-   - Angular CLI (install with: npm install -g @angular/cli)
-
-2. Setup:
-   - Windows: Double-click setup.bat
-   - Linux/Mac: Run ./setup.sh (make it executable first: chmod +x setup.sh)
-   - Manual: Run 'npm install' in the project directory
-
-3. Development:
-   - Run: npm start (or ng serve)
-   - Open: http://localhost:4200
-
-4. Build for production:
-   - Run: npm run build
-   - Files will be in: dist/bdd-agentic-workflow/
-
-FEATURES:
-=========
-✓ Modern Angular 17 with standalone components
-✓ Angular Material Design components
-✓ Fully responsive design
-✓ TypeScript support
-✓ SCSS styling with design system
-✓ Job status management interface
-✓ Real-time status tracking
-✓ Interactive job grid
-✓ Search and filter functionality
-
-CONVERSION NOTES:
-=================
-This project is a complete Angular conversion of the original React BDD Agentic Workflow application.
-
-Key changes made:
-- React components → Angular standalone components
-- JSX templates → Angular templates with directives
-- React hooks → Angular lifecycle methods
-- shadcn/ui → Angular Material
-- Lucide React → Material Icons
-- CSS modules → SCSS with utility classes
-
-The functionality and visual design remain identical to the original React version.
