--- angular/src/assets/.gitkeep
@@ -1,2 +1,0 @@
-/* Placeholder for assets */
-/* Add your images, fonts, and other static assets here */
