--- angular/src/favicon.ico
@@ -1,10 +1,0 @@
-<!doctype html>
-<html>
-<head>
-  <title>Favicon</title>
-</head>
-<body>
-  <!-- This is a placeholder favicon file -->
-  <!-- Replace this with your actual favicon.ico -->
-</body>
-</html>
