--- angular/src/main.ts
@@ -1,36 +1,0 @@
-import { bootstrapApplication } from '@angular/platform-browser';
-import { AppComponent } from './app/app.component';
-import { provideAnimationsAsync } from '@angular/platform-browser/animations/async';
-import { importProvidersFrom } from '@angular/core';
-import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule } from '@angular/material/card';
-import { MatInputModule } from '@angular/material/input';
-import { MatSelectModule } from '@angular/material/select';
-import { MatTableModule } from '@angular/material/table';
-import { MatChipsModule } from '@angular/material/chips';
-import { MatIconModule } from '@angular/material/icon';
-import { MatDividerModule } from '@angular/material/divider';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
-import { MatSnackBarModule } from '@angular/material/snack-bar';
-import { HttpClientModule } from '@angular/common/http';
-
-bootstrapApplication(AppComponent, {
-  providers: [
-    provideAnimationsAsync(),
-    importProvidersFrom(
-      HttpClientModule,
-      MatButtonModule,
-      MatCardModule,
-      MatInputModule,
-      MatSelectModule,
-      MatTableModule,
-      MatChipsModule,
-      MatIconModule,
-      MatDividerModule,
-      MatFormFieldModule,
-      MatProgressSpinnerModule,
-      MatSnackBarModule
-    )
-  ]
-}).catch(err => console.error(err));
