--- angular/src/styles.scss
@@ -1,118 +1,0 @@
-/* You can add global styles to this file, and also import other style files */
-@import '@angular/material/prebuilt-themes/indigo-pink.css';
-
-:root {
-  --font-size: 16px;
-  --background: #ffffff;
-  --foreground: #0a0a0a;
-  --card: #ffffff;
-  --card-foreground: #0a0a0a;
-  --popover: #ffffff;
-  --popover-foreground: #0a0a0a;
-  --primary: #030213;
-  --primary-foreground: #ffffff;
-  --secondary: #f5f5f5;
-  --secondary-foreground: #030213;
-  --muted: #ececf0;
-  --muted-foreground: #717182;
-  --accent: #e9ebef;
-  --accent-foreground: #030213;
-  --destructive: #d4183d;
-  --destructive-foreground: #ffffff;
-  --border: rgba(0, 0, 0, 0.1);
-  --input: transparent;
-  --input-background: #f3f3f5;
-  --switch-background: #cbced4;
-  --font-weight-medium: 500;
-  --font-weight-normal: 400;
-  --ring: #708090;
-  --radius: 0.625rem;
-}
-
-.dark {
-  --background: #0a0a0a;
-  --foreground: #ffffff;
-  --card: #0a0a0a;
-  --card-foreground: #ffffff;
-  --popover: #0a0a0a;
-  --popover-foreground: #ffffff;
-  --primary: #ffffff;
-  --primary-foreground: #0a0a0a;
-  --secondary: #1a1a1a;
-  --secondary-foreground: #ffffff;
-  --muted: #1a1a1a;
-  --muted-foreground: #8a8a8a;
-  --accent: #1a1a1a;
-  --accent-foreground: #ffffff;
-  --destructive: #dc2626;
-  --destructive-foreground: #ffffff;
-  --border: #1a1a1a;
-  --input: #1a1a1a;
-  --ring: #6b7280;
-}
-
-html, body {
-  height: 100%;
-  margin: 0;
-  padding: 0;
-  font-family: 'Roboto', sans-serif;
-  background-color: var(--background);
-  color: var(--foreground);
-}
-
-/* Custom utility classes */
-.text-2xl {
-  font-size: 1.5rem;
-  line-height: 2rem;
-  font-weight: 600;
-}
-
-.text-muted-foreground {
-  color: var(--muted-foreground);
-}
-
-.bg-background {
-  background-color: var(--background);
-}
-
-.bg-card {
-  background-color: var(--card);
-}
-
-.border-b {
-  border-bottom: 1px solid var(--border);
-}
-
-/* Material Design customizations */
-.mat-mdc-card {
-  background-color: var(--card) !important;
-  color: var(--card-foreground) !important;
-  border: 1px solid var(--border);
-  border-radius: var(--radius);
-}
-
-.mat-mdc-button {
-  border-radius: calc(var(--radius) - 2px);
-}
-
-.mat-mdc-form-field {
-  .mat-mdc-text-field-wrapper {
-    border-radius: calc(var(--radius) - 2px);
-  }
-}
-
-.mat-mdc-table {
-  background-color: transparent !important;
-  
-  .mat-mdc-header-row {
-    background-color: var(--muted);
-  }
-  
-  .mat-mdc-row:hover {
-    background-color: var(--accent);
-  }
-}
-
-.mat-mdc-chip {
-  border-radius: calc(var(--radius) - 2px);
-}
