+++ frontend/.angular/cache/17.3.17/babel-webpack/57a07852a2e13bf76e7d3b26e753b338c5b879302c59ac4cc430d56d260292a7.json
@@ -1,0 +1,1 @@
+{"ast":null,"code":"import { __decorate } from \"tslib\";\nimport { Component } from '@angular/core';\nimport { CommonModule } from '@angular/common';\nimport { MatButtonModule } from '@angular/material/button';\nimport { MatChipsModule } from '@angular/material/chips';\nimport { MatIconModule } from '@angular/material/icon';\nlet StudyHeaderComponent = class StudyHeaderComponent {};\nStudyHeaderComponent = __decorate([Component({\n  selector: 'app-study-header',\n  standalone: true,\n  imports: [CommonModule, MatButtonModule, MatChipsModule, MatIconModule],\n  template: `\n    <div class=\"border-b bg-card px-6 py-4\">\n      <div class=\"flex items-center justify-between\">\n        <div class=\"space-y-1\">\n          <h1 class=\"text-2xl\">BDD Agentic Workflow</h1>\n          <p class=\"text-muted-foreground\">Behavior-driven development with intelligent automation</p>\n        </div>\n        <div class=\"flex items-center gap-3\">\n\n          <button mat-stroked-button>\n            <mat-icon>settings</mat-icon>\n            Settings\n          </button>\n        </div>\n      </div>\n    </div>\n  `,\n  styles: [`\n    .border-b { border-bottom: 1px solid var(--border, rgba(0, 0, 0, 0.1)); }\n    .bg-card { background-color: var(--card, #ffffff); }\n    .px-6 { padding-left: 1.5rem; padding-right: 1.5rem; }\n    .py-4 { padding-top: 1rem; padding-bottom: 1rem; }\n    .flex { display: flex; }\n    .items-center { align-items: center; }\n    .justify-between { justify-content: space-between; }\n    .space-y-1 > * + * { margin-top: 0.25rem; }\n    .text-2xl { font-size: 1.5rem; line-height: 2rem; }\n    .text-muted-foreground { color: var(--muted-foreground, #717182); }\n    .gap-3 { gap: 0.75rem; }\n    .px-3 { padding-left: 0.75rem; padding-right: 0.75rem; }\n    .py-1 { padding-top: 0.25rem; padding-bottom: 0.25rem; }\n    mat-icon { margin-right: 0.5rem; }\n  `]\n})], StudyHeaderComponent);\nexport { StudyHeaderComponent };","map":{"version":3,"names":["Component","CommonModule","MatButtonModule","MatChipsModule","MatIconModule","StudyHeaderComponent","__decorate","selector","standalone","imports","template","styles"],"sources":["F:\\aaa\\jira_2_bdd_agentic_workflow\\angular\\src\\app\\components\\study-header\\study-header.component.ts"],"sourcesContent":["import { Component } from '@angular/core';\nimport { CommonModule } from '@angular/common';\nimport { MatButtonModule } from '@angular/material/button';\nimport { MatChipsModule } from '@angular/material/chips';\nimport { MatIconModule } from '@angular/material/icon';\n\n@Component({\n  selector: 'app-study-header',\n  standalone: true,\n  imports: [CommonModule, MatButtonModule, MatChipsModule, MatIconModule],\n  template: `\n    <div class=\"border-b bg-card px-6 py-4\">\n      <div class=\"flex items-center justify-between\">\n        <div class=\"space-y-1\">\n          <h1 class=\"text-2xl\">BDD Agentic Workflow</h1>\n          <p class=\"text-muted-foreground\">Behavior-driven development with intelligent automation</p>\n        </div>\n        <div class=\"flex items-center gap-3\">\n\n          <button mat-stroked-button>\n            <mat-icon>settings</mat-icon>\n            Settings\n          </button>\n        </div>\n      </div>\n    </div>\n  `,\n  styles: [`\n    .border-b { border-bottom: 1px solid var(--border, rgba(0, 0, 0, 0.1)); }\n    .bg-card { background-color: var(--card, #ffffff); }\n    .px-6 { padding-left: 1.5rem; padding-right: 1.5rem; }\n    .py-4 { padding-top: 1rem; padding-bottom: 1rem; }\n    .flex { display: flex; }\n    .items-center { align-items: center; }\n    .justify-between { justify-content: space-between; }\n    .space-y-1 > * + * { margin-top: 0.25rem; }\n    .text-2xl { font-size: 1.5rem; line-height: 2rem; }\n    .text-muted-foreground { color: var(--muted-foreground, #717182); }\n    .gap-3 { gap: 0.75rem; }\n    .px-3 { padding-left: 0.75rem; padding-right: 0.75rem; }\n    .py-1 { padding-top: 0.25rem; padding-bottom: 0.25rem; }\n    mat-icon { margin-right: 0.5rem; }\n  `]\n})\nexport class StudyHeaderComponent {}"],"mappings":";AAAA,SAASA,SAAS,QAAQ,eAAe;AACzC,SAASC,YAAY,QAAQ,iBAAiB;AAC9C,SAASC,eAAe,QAAQ,0BAA0B;AAC1D,SAASC,cAAc,QAAQ,yBAAyB;AACxD,SAASC,aAAa,QAAQ,wBAAwB;AAwC/C,IAAMC,oBAAoB,GAA1B,MAAMA,oBAAoB,GAAG;AAAvBA,oBAAoB,GAAAC,UAAA,EAtChCN,SAAS,CAAC;EACTO,QAAQ,EAAE,kBAAkB;EAC5BC,UAAU,EAAE,IAAI;EAChBC,OAAO,EAAE,CAACR,YAAY,EAAEC,eAAe,EAAEC,cAAc,EAAEC,aAAa,CAAC;EACvEM,QAAQ,EAAE;;;;;;;;;;;;;;;;GAgBT;EACDC,MAAM,EAAE,CAAC;;;;;;;;;;;;;;;GAeR;CACF,CAAC,C,EACWN,oBAAoB,CAAG","ignoreList":[]},"metadata":{},"sourceType":"module","externalDependencies":[]}
