+++ frontend/src/assets/.gitkeep
@@ -1,0 +1,2 @@
+/* Placeholder for assets */
+/* Add your images, fonts, and other static assets here */
